Revert "Merge branch 'LF-15763-missions-ended-generic-indicator-for-missions' of https://github.com/jumperexchange/jumper-exchange into LF-15763-missions-ended-generic-indicator-for-missions"
This reverts commit 8748f9a830918525ab9f340f47be082e5805823d, reversing
changes made to ce4e0690d3b03f68b781d248e3ba4dc3281461f8.
