+++ tests/testData/connectWalletFunctions.ts
@@ -1,5 +1,7 @@
-import type { Page } from '@playwright/test';
+import type { Page, BrowserContext } from '@playwright/test';
 import { expect } from '@playwright/test';
+import { MetaMask } from '@synthetixio/synpress/playwright';
+import basicSetup from '../wallet-setup/basic.setup';
 
 export const getRankValue = async (page: Page) => {
   const rankSelector = 'xpath=(//div[@class="MuiBox-root mui-19kp780"]//p)[1]';
@@ -34,5 +36,35 @@ export const connectAnotherWalletButton = (page: Page) => {
 
 export const disconnectWalletButton = (page: Page) => {
   return page.locator('#disconnect-wallet-button');
+};
+
+/**
+ * Reusable function to connect MetaMask wallet to the application
+ * @param context - Browser context
+ * @param page - Playwright page
+ * @param extensionId - MetaMask extension ID
+ * @param targetUrl - URL to navigate to (defaults to '/')
+ */
+export const connectMetaMaskWallet = async (
+  context: BrowserContext,
+  page: Page,
+  extensionId: string,
+  targetUrl: string = '/'
+) => {
+  const metamask = new MetaMask(
+    context,
+    page,
+    basicSetup.walletPassword,
+    extensionId,
+  );
+  const metaMaskWalletOption = page.locator(
+    'xpath=//span[normalize-space(text())="MetaMask"]',
+  );
+  
+  await page.goto(targetUrl);
+  await expect(connectButton(page)).toBeEnabled();
+  await connectButton(page).click();
+  await metaMaskWalletOption.click();
+  await metamask.connectToDapp(['Account 1']);
 };
